Templates for api integration
