+++ lib/core/error/failure.dart
@@ -1,3 +1,1 @@
-class Failure {
-  Failure([List properties = const <dynamic>[]]) : super(properties);
-}
+class Failure {}
