+++ lib/features/auth/forgot_password/data/models/forgot_password_post_profile.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/foundation.dart';
+
+//Temporary holder for api change it acordingly when docs get updated
+
+class ForgotPasswordModel {
+  final int number;
+  ForgotPasswordModel({
+    @required this.number,
+  });
+
+  Map<String, int> toJson() => {'phoneNumber': number};
+}
