+++ lib/features/auth/forgot_password/data/models/otp_reciever_profile.dart
@@ -1,0 +1,1 @@
+class OtpReciever {}
